Develop (#148)
* Ben update alerts (#145)

* chore: copy changes to data alerts modal

AC-636

* cleanup

* fix: toolbar overlap issue

There was an issue where the display types would overlap the menu options if there were too many
display types

AC-898

* ..

* chore: remove create data alert from notifications tab inside DM

Created a prop for the create data alert button so that it is hidden by default, but can be added if
desired. This is because creating data alerts won't be intuitive to many non-technical users

AC-815

* feat(datamessenger): support OAuth2 for ms-dynamics

Support authentication in microsoft for querying from their environment

AC-565

* Ben update alerts (#146)

* chore: copy changes to data alerts modal

AC-636

* fix: toolbar overlap issue

There was an issue where the display types would overlap the menu options if there were too many
display types

AC-898

* chore: remove create data alert from notifications tab inside DM

Created a prop for the create data alert button so that it is hidden by default, but can be added if
desired. This is because creating data alerts won't be intuitive to many non-technical users

AC-815

* feat(datamessenger): support OAuth2 for ms-dynamics

Support authentication in microsoft for querying from their environment

AC-565

* fix: charting fix for multiseries lists

Fixed an issue with multi-series list queries not showing supported chart type options

AC-926

* chore: updates to data alerts modal

Copy and minor functional updates to the create data alert flow

* fix: data messenger styling bug

* fix(dashboard.js): aNT-647

Co-authored-by: AlexLiu-Chata <alui@chata.ai>
Co-authored-by: AlexLiu-Chata <79466486+AlexLiu-Chata@users.noreply.github.com>